Track selection state of module within module itself
This change removes the selected `ComponentState` instance from both the `SelectorState` and `EdgeState` implementations.
The currently selected component is now owned solely by the `ModuleResolveState`.
Each `SelectorState` can either have a failure, or bind to a successfully resolved module.
Likewise, an `EdgeState` will either be linked to a `SelectorState` with a failure,
or will have a target component from the resolved module.